+++ journal.go
@@ -3,9 +3,12 @@ package concator
 import (
 	"fmt"
 	"io"
+	"io/ioutil"
+	"path/filepath"
 	"sync"
-	"sync/atomic"
 	"time"
+
+	"github.com/pkg/errors"
 
 	"github.com/Laisky/go-fluentd/libs"
 	"github.com/Laisky/go-fluentd/monitor"
@@ -14,10 +17,19 @@ import (
 	"github.com/Laisky/zap"
 )
 
+const (
+	defaultInnerJournalDataChanLen = 1000
+	defaultInnerJournalIDChanLen   = 1000
+	minimalBufSizeByte             = 10485760 // 10 MB
+	intervalToStartingLegacy       = 3 * time.Second
+	intervalForceGC                = 1 * time.Minute
+)
+
 type JournalCfg struct {
 	BufDirPath                         string
 	BufSizeBytes                       int64
 	JournalOutChanLen, CommitIdChanLen int
+	IsCompress                         bool
 	MsgPool                            *sync.Pool
 }
 
@@ -25,9 +37,17 @@ type JournalCfg struct {
 // then check every msg with committed id to make sure no msg lost
 type Journal struct {
 	*JournalCfg
-	j          *journal.Journal
-	legacyLock uint32
+	legacyLock *utils.Mutex
+
 	outChan    chan *libs.FluentMsg
+	commitChan chan *libs.FluentMsg
+
+	baseJournalDir      string
+	baseJournalCfg      *journal.JournalConfig
+	jjLock              *sync.Mutex
+	tag2JMap            *sync.Map // map[string]*journal.Journal
+	tag2JJInchanMap     *sync.Map // map[string]chan *libs.FluentMsg
+	tag2JJCommitChanMap *sync.Map // map[string]chan *libs.FluentMsg
 }
 
 // NewJournal create new Journal with `bufDirPath` and `BufSizeBytes`
@@ -35,27 +55,289 @@ func NewJournal(cfg *JournalCfg) *Journal {
 	utils.Logger.Info("create new journal",
 		zap.String("filepath", cfg.BufDirPath),
 		zap.Int64("size", cfg.BufSizeBytes))
-	if cfg.BufSizeBytes < 10485760 { // 10MB
+	if cfg.BufSizeBytes < minimalBufSizeByte {
 		utils.Logger.Warn("journal buf file size too small", zap.Int64("size", cfg.BufSizeBytes))
 	}
 
 	jcfg := journal.NewConfig()
 	jcfg.BufDirPath = cfg.BufDirPath
 	jcfg.BufSizeBytes = cfg.BufSizeBytes
+	jcfg.IsCompress = cfg.IsCompress
+	// jcfg.RotateDuration = 3 * time.Second // TODO
 
 	j := &Journal{
 		JournalCfg: cfg,
-		j:          journal.NewJournal(jcfg),
-		legacyLock: 0,
+		legacyLock: &utils.Mutex{},
+
+		commitChan: make(chan *libs.FluentMsg, cfg.CommitIdChanLen),
 		outChan:    make(chan *libs.FluentMsg, cfg.JournalOutChanLen),
+
+		jjLock:              &sync.Mutex{},
+		baseJournalDir:      jcfg.BufDirPath,
+		baseJournalCfg:      jcfg,
+		tag2JMap:            &sync.Map{},
+		tag2JJInchanMap:     &sync.Map{},
+		tag2JJCommitChanMap: &sync.Map{},
 	}
+	j.initLegacyJJ()
 	j.registerMonitor()
+	j.startCommitRunner()
 	return j
 }
 
+// initLegacyJJ process existed legacy data and ids
+func (j *Journal) initLegacyJJ() {
+	files, err := ioutil.ReadDir(j.baseJournalDir)
+	if err != nil {
+		utils.Logger.Warn("try to read dir of journal got error", zap.Error(err))
+		return
+	}
+
+	for _, dir := range files {
+		if dir.IsDir() {
+			j.createJournalRunner(dir.Name())
+		}
+	}
+}
+
 // LoadMaxID load the max committed id from journal
-func (j *Journal) LoadMaxID() (int64, error) {
-	return j.j.LoadMaxId()
+func (j *Journal) LoadMaxID() (id int64, err error) {
+	var (
+		nid  int64
+		tag  string
+		jj   *journal.Journal
+		err2 error
+	)
+	j.tag2JMap.Range(func(k, v interface{}) bool {
+		tag = k.(string)
+		jj = v.(*journal.Journal)
+		if nid, err2 = jj.LoadMaxId(); err2 != nil {
+			if nid > id {
+				id = nid
+			}
+		} else {
+			err = errors.Wrapf(err2, "try to load max id with tag `%v` got error", tag)
+		}
+
+		return true
+	})
+
+	return nid, err
+}
+
+func (j *Journal) ProcessLegacyMsg(dumpChan chan *libs.FluentMsg) (maxID int64, err2 error) {
+	if !j.legacyLock.TryLock() {
+		return 0, fmt.Errorf("another legacy is running")
+	}
+	defer j.legacyLock.ForceRealse()
+
+	utils.Logger.Debug("starting to process legacy data...")
+	var (
+		wg = &sync.WaitGroup{}
+		l  = &sync.Mutex{}
+	)
+
+	j.tag2JMap.Range(func(k, v interface{}) bool {
+		wg.Add(1)
+		go func(tag string, jj *journal.Journal) {
+			defer wg.Done()
+			var (
+				innerMaxID int64
+				err        error
+				msg        *libs.FluentMsg
+				data       = &journal.Data{Data: map[string]interface{}{}}
+			)
+
+			if !jj.LockLegacy() { // avoid rotate
+				return
+			}
+
+			startTs := utils.Clock.GetUTCNow()
+			for {
+				msg = j.MsgPool.Get().(*libs.FluentMsg)
+				data.Data["message"] = nil // alloc new map to avoid old data contaminate
+				if err = jj.LoadLegacyBuf(data); err == io.EOF {
+					utils.Logger.Debug("load legacy buf done",
+						zap.Float64("sec", utils.Clock.GetUTCNow().Sub(startTs).Seconds()),
+					)
+					j.MsgPool.Put(msg)
+
+					l.Lock()
+					if innerMaxID > maxID {
+						maxID = innerMaxID
+					}
+					l.Unlock()
+					return
+				} else if err != nil {
+					utils.Logger.Error("load legacy data got error", zap.Error(err))
+					j.MsgPool.Put(msg)
+					if !jj.LockLegacy() {
+						l.Lock()
+						if innerMaxID > maxID {
+							maxID = innerMaxID
+						}
+						err2 = err
+						l.Unlock()
+						return
+					}
+					continue
+				}
+
+				if data.Data["message"] == nil {
+					utils.Logger.Warn("lost message")
+					j.MsgPool.Put(msg)
+					continue
+				}
+
+				msg.Id = data.ID
+				msg.Tag = string(data.Data["tag"].(string))
+				msg.Message = data.Data["message"].(map[string]interface{})
+				if msg.Id > innerMaxID {
+					innerMaxID = msg.Id
+				}
+				utils.Logger.Debug("load msg from legacy",
+					zap.String("tag", msg.Tag),
+					zap.Int64("id", msg.Id))
+
+				// rewrite data into journal
+				// only committed id can really remove a msg
+				dumpChan <- msg
+			}
+		}(k.(string), v.(*journal.Journal))
+
+		return true
+	})
+
+	wg.Wait()
+	utils.Logger.Debug("process legacy done")
+	return
+}
+
+// createJournalRunner create journal for a tag,
+// and return commit channel and dump channel
+func (j *Journal) createJournalRunner(tag string) {
+	j.jjLock.Lock()
+	defer j.jjLock.Unlock()
+
+	var ok bool
+	if _, ok = j.tag2JMap.Load(tag); ok {
+		return // double check to prevent duplicate create jj runner
+	}
+
+	jcfg := journal.NewConfig()
+	jcfg.BufDirPath = j.baseJournalCfg.BufDirPath
+	jcfg.BufSizeBytes = j.baseJournalCfg.BufSizeBytes
+	jcfg.IsCompress = j.baseJournalCfg.IsCompress
+	jcfg.IsAggresiveGC = false
+	jcfg.BufDirPath = filepath.Join(j.baseJournalDir, tag)
+
+	utils.Logger.Info("createJournalRunner for tag", zap.String("tag", tag))
+	if _, ok = j.tag2JMap.Load(tag); ok {
+		utils.Logger.Panic("tag already exists in tag2JMap", zap.String("tag", tag))
+	}
+	utils.Logger.Info("create new journal.Journal", zap.String("tag", tag))
+	jj := journal.NewJournal(jcfg)
+	j.tag2JMap.Store(tag, jj)
+
+	if _, ok = j.tag2JJInchanMap.Load(tag); ok {
+		utils.Logger.Panic("tag already exists in tag2JJInchanMap", zap.String("tag", tag))
+	}
+	j.tag2JJInchanMap.Store(tag, make(chan *libs.FluentMsg, defaultInnerJournalDataChanLen))
+
+	if _, ok = j.tag2JJCommitChanMap.Load(tag); ok {
+		utils.Logger.Panic("tag already exists in tag2JJCommitChanMap", zap.String("tag", tag))
+	}
+	j.tag2JJCommitChanMap.Store(tag, make(chan *libs.FluentMsg, defaultInnerJournalIDChanLen))
+
+	// create ids writer
+	go func() {
+		defer utils.Logger.Panic("journal ids writer quit")
+		var (
+			mid      int64
+			err      error
+			nRetry   = 0
+			maxRetry = 2
+		)
+
+		chani, ok := j.tag2JJCommitChanMap.Load(tag)
+		if !ok {
+			utils.Logger.Panic("tag must in `j.tag2JJCommitChanMap`", zap.String("tag", tag))
+		}
+
+		for msg := range chani.(chan *libs.FluentMsg) {
+			nRetry = 0
+			for nRetry < maxRetry {
+				if err = jj.WriteId(msg.Id); err != nil {
+					nRetry++
+				}
+				break
+			}
+			if err != nil && nRetry == maxRetry {
+				utils.Logger.Error("try to write id to journal got error", zap.Error(err))
+			}
+
+			if msg.ExtIds != nil {
+				for _, mid = range msg.ExtIds {
+					nRetry = 0
+					for nRetry < maxRetry {
+						if err = jj.WriteId(mid); err != nil {
+							nRetry++
+						}
+						break
+					}
+					if err != nil && nRetry == maxRetry {
+						utils.Logger.Error("try to write id to journal got error", zap.Error(err))
+					}
+				}
+				msg.ExtIds = nil
+			}
+
+			j.MsgPool.Put(msg)
+		}
+	}()
+
+	// create data writer
+	go func() {
+		defer utils.Logger.Panic("journal data writer quit")
+
+		var (
+			data     = &journal.Data{Data: map[string]interface{}{}}
+			err      error
+			nRetry   = 0
+			maxRetry = 2
+		)
+		chani, ok := j.tag2JJInchanMap.Load(tag)
+		if !ok {
+			utils.Logger.Panic("tag should in `j.tag2JJInchanMap`", zap.String("tag", tag))
+		}
+
+		for msg := range chani.(chan *libs.FluentMsg) {
+			data.ID = msg.Id
+			data.Data["message"] = msg.Message
+			data.Data["tag"] = msg.Tag
+			nRetry = 0
+			for nRetry < maxRetry {
+				if err = jj.WriteData(data); err != nil {
+					nRetry++
+				}
+				break
+			}
+			if err != nil && nRetry == maxRetry {
+				utils.Logger.Error("try to write msg to journal got error",
+					zap.Error(err),
+					zap.String("tag", msg.Tag),
+				)
+			}
+
+			select {
+			case j.outChan <- msg:
+			default:
+				// msg will reproduce in legacy stage,
+				// so you can discard msg without any side-effect.
+				j.MsgPool.Put(msg)
+			}
+		}
+	}()
 }
 
 func (j *Journal) GetOutChan() chan *libs.FluentMsg {
@@ -68,79 +350,29 @@ func (j *Journal) ConvertMsg2Buf(msg *libs.FluentMsg, data *map[string]interface
 	(*data)["message"] = msg.Message
 }
 
-func (j *Journal) ProcessLegacyMsg(msgPool *sync.Pool, msgChan, dumpChan chan *libs.FluentMsg) (maxId int64, err error) {
-	utils.Logger.Debug("starting to process legacy data...")
-	if ok := atomic.CompareAndSwapUint32(&j.legacyLock, 0, 1); !ok {
-		utils.Logger.Debug("legacy data already in processing")
-		return 0, nil
-	}
-	defer atomic.StoreUint32(&j.legacyLock, 0)
-
-	var (
-		i    = 0
-		msg  *libs.FluentMsg
-		data = &journal.Data{Data: map[string]interface{}{}}
-	)
-
-	if !j.j.LockLegacy() { // avoid rotate
-		return
-	}
-
-	startTs := utils.Clock.GetUTCNow()
-	for {
-		i++
-		msg = msgPool.Get().(*libs.FluentMsg)
-		// utils.Logger.Info(fmt.Sprintf("got %p", msg))
-		data.Data["message"] = nil // alloc new map to avoid old data contaminate
-
-		if err = j.j.LoadLegacyBuf(data); err == io.EOF {
-			utils.Logger.Debug("load legacy buf done", zap.Int("n", i), zap.Float64("sec", utils.Clock.GetUTCNow().Sub(startTs).Seconds()))
-			msgPool.Put(msg)
-			// utils.Logger.Info(fmt.Sprintf("recycle: %p", msg))
-			return maxId, nil
-		} else if err != nil {
-			utils.Logger.Error("load legacy data got error", zap.Error(err))
-			msgPool.Put(msg)
-			// utils.Logger.Info(fmt.Sprintf("recycle: %p", msg))
-			continue
-		}
-
-		if data.Data["message"] == nil {
-			utils.Logger.Warn("lost message")
-			msgPool.Put(msg)
-			// utils.Logger.Info(fmt.Sprintf("recycle: %p, %v", msg))
-			continue
-		}
-
-		msg.Id = data.ID
-		msg.Tag = string(data.Data["tag"].(string))
-		msg.Message = data.Data["message"].(map[string]interface{})
-		if msg.Id > maxId {
-			maxId = msg.Id
-		}
-
-		// rewrite data into journal
-		// only committed id can really remove a msg
-		dumpChan <- msg
-	}
-}
-
-// IntervalToStartingLegacy interval to restarting legacy loading
-const IntervalToStartingLegacy = 3 * time.Second // arbitary
-
 func (j *Journal) DumpMsgFlow(msgPool *sync.Pool, dumpChan, skipDumpChan chan *libs.FluentMsg) chan *libs.FluentMsg {
 	// deal with legacy
 	go func() {
 		defer utils.Logger.Panic("legacy processor exit")
 		var err error
 		for { // try to starting legacy loading
-			if _, err = j.ProcessLegacyMsg(msgPool, j.outChan, dumpChan); err != nil {
+			if _, err = j.ProcessLegacyMsg(dumpChan); err != nil {
 				utils.Logger.Error("process legacy got error", zap.Error(err))
 			}
-			time.Sleep(IntervalToStartingLegacy)
+			time.Sleep(intervalToStartingLegacy)
 		}
 	}()
 
+	// start periodic gc
+	go func() {
+		defer utils.Logger.Panic("gc runner exit")
+		for {
+			utils.ForceGC()
+			time.Sleep(intervalForceGC)
+		}
+	}()
+
+	// deal with msgs that skip dump
 	go func() {
 		defer utils.Logger.Panic("skipDumpChan goroutine exit")
 		for msg := range skipDumpChan {
@@ -149,42 +381,27 @@ func (j *Journal) DumpMsgFlow(msgPool *sync.Pool, dumpChan, skipDumpChan chan *l
 	}()
 
 	go func() {
+		defer utils.Logger.Panic("legacy dumper exit")
+
 		var (
-			err      error
-			data     = &journal.Data{Data: map[string]interface{}{}}
-			nRetry   = 0
-			maxRetry = 5
-			msg      *libs.FluentMsg
+			ok  bool
+			jji interface{}
 		)
-		defer utils.Logger.Panic("legacy dumper exit", zap.String("msg", fmt.Sprint(msg)))
-
-		for msg = range dumpChan {
-			data.ID = msg.Id
-			data.Data["message"] = msg.Message
-			data.Data["tag"] = msg.Tag
-			// utils.Logger.Debug("got new message", zap.Int64("id", msg.Id), zap.String("tag", msg.Tag))
-			for {
-				if err = j.j.WriteData(data); err != nil {
-					nRetry++
-					if nRetry < maxRetry {
-						utils.Logger.Warn("try to write data got error", zap.String("tag", msg.Tag), zap.Int64("id", msg.Id), zap.Error(err))
-						continue
-					}
-
-					nRetry = 0
-					utils.Logger.Error("try to write data got error", zap.String("tag", msg.Tag), zap.Int64("id", msg.Id), zap.Error(err))
-					break
-				}
-
-				break
+		for msg := range dumpChan {
+			utils.Logger.Debug("try to dump msg", zap.String("tag", msg.Tag))
+			if jji, ok = j.tag2JJInchanMap.Load(msg.Tag); !ok {
+				j.createJournalRunner(msg.Tag)
+				jji, _ = j.tag2JJInchanMap.Load(msg.Tag)
 			}
 
-			// utils.Logger.Debug("success write data journal", zap.String("tag", msg.Tag), zap.Int64("id", msg.Id))
-
 			select {
+			case jji.(chan *libs.FluentMsg) <- msg:
 			case j.outChan <- msg:
 			default:
-				// utils.Logger.Info(fmt.Sprintf("recycle: %p, %v", msg))
+				utils.Logger.Error("discard msg since of journal & downstream busy",
+					zap.String("tag", msg.Tag),
+					zap.String("msg", fmt.Sprint(msg)),
+				)
 				j.MsgPool.Put(msg)
 			}
 		}
@@ -193,40 +410,47 @@ func (j *Journal) DumpMsgFlow(msgPool *sync.Pool, dumpChan, skipDumpChan chan *l
 	return j.outChan
 }
 
-func (j *Journal) GetCommitChan() chan<- int64 {
-	var (
-		err        error
-		nRetry     = 0
-		maxRetry   = 5
-		commitChan = make(chan int64, j.CommitIdChanLen)
-	)
+func (j *Journal) GetCommitChan() chan<- *libs.FluentMsg {
+	return j.commitChan
+}
+
+func (j *Journal) startCommitRunner() {
 	go func() {
 		defer utils.Logger.Panic("id commitor exit")
 
-		for id := range commitChan {
-			for {
-				if err = j.j.WriteId(id); err != nil {
-					utils.Logger.Warn("write id got error", zap.Int64("id", id), zap.Error(err))
-					nRetry++
-					if nRetry < maxRetry {
-						continue
-					}
+		var (
+			ok    bool
+			chani interface{}
+		)
+		for msg := range j.commitChan {
+			utils.Logger.Debug("try to commit msg",
+				zap.String("tag", msg.Tag),
+				zap.Int64("id", msg.Id))
+			if chani, ok = j.tag2JJCommitChanMap.Load(msg.Tag); !ok {
+				j.createJournalRunner(msg.Tag)
+				chani, _ = j.tag2JJCommitChanMap.Load(msg.Tag)
+			}
 
-					nRetry = 0
-					utils.Logger.Error("try to write id got error", zap.Int64("id", id))
-					break
-				}
-
-				break
+			select {
+			case chani.(chan *libs.FluentMsg) <- msg:
+			default:
+				utils.Logger.Error("discard id without commit since of journal is busy",
+					zap.String("tag", msg.Tag),
+					zap.Int64("id", msg.Id),
+				)
+				j.MsgPool.Put(msg)
 			}
 		}
 	}()
-
-	return commitChan
 }
 
 func (j *Journal) registerMonitor() {
 	monitor.AddMetric("journal", func() map[string]interface{} {
-		return j.j.GetMetric()
+		result := map[string]interface{}{}
+		j.tag2JMap.Range(func(k, v interface{}) bool {
+			result[k.(string)] = v.(*journal.Journal).GetMetric()
+			return true
+		})
+		return result
 	})
 }