+++ recvs/fluentd.go
@@ -5,6 +5,9 @@ import (
 	"fmt"
 	"io"
 	"net"
+	"regexp"
+	"sync"
+	"time"
 
 	"github.com/Laisky/go-fluentd/libs"
 	utils "github.com/Laisky/go-utils"
@@ -20,21 +23,59 @@ type FluentdRecvCfg struct {
 
 	// if IsRewriteTagFromTagKey, set `msg.Tag = msg.Message[TagKey]`
 	IsRewriteTagFromTagKey bool
+
+	ConcatMaxLen int
+	ConcatCfg    map[string]interface{}
+}
+
+type concatCfg struct {
+	headRegexp *regexp.Regexp
+	msgKey,
+	identifierKey string
 }
 
 // FluentdRecv recv for fluentd format
 type FluentdRecv struct {
 	*BaseRecv
 	*FluentdRecvCfg
+
+	concatTagCfg   map[string]*concatCfg
+	pendingMsgPool *sync.Pool
+}
+
+// PendingMsg is the message wait tobe concatenate
+type PendingMsg struct {
+	msg   *libs.FluentMsg
+	lastT time.Time
 }
 
 // NewFluentdRecv create new FluentdRecv
 func NewFluentdRecv(cfg *FluentdRecvCfg) *FluentdRecv {
 	utils.Logger.Info("create FluentdRecv")
-	return &FluentdRecv{
+	r := &FluentdRecv{
 		BaseRecv:       &BaseRecv{},
 		FluentdRecvCfg: cfg,
+		pendingMsgPool: &sync.Pool{
+			New: func() interface{} {
+				return &PendingMsg{}
+			},
+		},
+		concatTagCfg: map[string]*concatCfg{},
 	}
+
+	tags := []string{}
+	for tag, cfgi := range cfg.ConcatCfg {
+		tags = append(tags, tag)
+		cfg := cfgi.(map[string]interface{})
+		r.concatTagCfg[tag] = &concatCfg{
+			identifierKey: cfg["identifier"].(string),
+			msgKey:        cfg["msg_key"].(string),
+			headRegexp:    regexp.MustCompile(cfg["head_regexp"].(string)),
+		}
+	}
+	utils.Logger.Info("enable concator for tags", zap.Strings("tags", tags))
+
+	return r
 }
 
 // GetName return the name of this recv
@@ -83,15 +124,18 @@ func (r *FluentdRecv) decodeMsg(conn net.Conn) {
 		tag     string
 		ok      bool
 		entryI  interface{}
+		conCtx  *concatCtx
 		eof     = msgp.WrapError(io.EOF)
 	)
 
 	for {
 		if err = v.DecodeMsg(reader); err == eof {
 			utils.Logger.Info("connection closed", zap.String("remote", conn.RemoteAddr().String()))
+			r.closeConcatCtx(conCtx)
 			return
 		} else if err != nil {
 			utils.Logger.Error("decode message got error", zap.Error(err))
+			r.closeConcatCtx(conCtx)
 			return
 		}
 
@@ -164,6 +208,7 @@ func (r *FluentdRecv) decodeMsg(conn net.Conn) {
 			}
 
 			utils.Logger.Debug("got message in format: default")
+
 			switch msgBody := v[2].(type) {
 			case map[string]interface{}:
 				msg = r.msgPool.Get().(*libs.FluentMsg)
@@ -174,6 +219,20 @@ func (r *FluentdRecv) decodeMsg(conn net.Conn) {
 			}
 
 			msg.Tag = tag
+			if _, ok = r.concatTagCfg[tag]; ok { // need concat
+				if conCtx == nil {
+					utils.Logger.Info("create new concatCtx for new conn")
+					conCtx = &concatCtx{
+						concatCfg:          r.concatTagCfg[tag],
+						identifier2LastMsg: map[string]*PendingMsg{},
+					}
+				}
+				msg = r.Concat(conCtx, msg)
+			}
+
+			if msg == nil {
+				continue
+			}
 			r.SendMsg(msg)
 		}
 	}
@@ -203,4 +262,108 @@ func (r *FluentdRecv) SendMsg(msg *libs.FluentMsg) {
 	msg.Id = r.counter.Count()
 	utils.Logger.Debug("receive new msg", zap.String("tag", msg.Tag), zap.Int64("id", msg.Id))
 	r.asyncOutChan <- msg
+}
+
+type concatCtx struct {
+	*concatCfg
+	identifier2LastMsg map[string]*PendingMsg
+
+	ok         bool
+	log        []byte
+	identifier string
+	pmsg       *PendingMsg
+}
+
+func (r *FluentdRecv) closeConcatCtx(ctx *concatCtx) {
+	if ctx == nil {
+		return
+	}
+
+	for _, ctx.pmsg = range ctx.identifier2LastMsg {
+		r.SendMsg(ctx.pmsg.msg)
+	}
+}
+
+// Concat concatenate with old messages
+func (r *FluentdRecv) Concat(ctx *concatCtx, msg *libs.FluentMsg) (newMsg *libs.FluentMsg) {
+	switch msg.Message[ctx.msgKey].(type) {
+	case []byte:
+		ctx.log = msg.Message[ctx.msgKey].([]byte)
+	case string:
+		ctx.log = []byte(msg.Message[ctx.msgKey].(string))
+		msg.Message[ctx.msgKey] = ctx.log
+	default:
+		utils.Logger.Warn("unknown msg key or unknown type",
+			zap.String("tag", msg.Tag),
+			zap.String("msg_key", ctx.msgKey),
+			zap.String("msg", fmt.Sprint(msg.Message)))
+		return msg
+	}
+
+	switch msg.Message[ctx.identifierKey].(type) {
+	case []byte:
+		ctx.identifier = string(msg.Message[ctx.identifierKey].([]byte))
+	case string:
+		ctx.identifier = msg.Message[ctx.identifierKey].(string)
+	default:
+		utils.Logger.Warn("unknown identifier or unknown type",
+			zap.String("tag", msg.Tag),
+			zap.String("identifier_key", ctx.identifier),
+			zap.String("identifier", fmt.Sprint(msg.Message[ctx.identifierKey])))
+		return msg
+	}
+
+	if ctx.pmsg, ctx.ok = ctx.identifier2LastMsg[ctx.identifier]; !ctx.ok { // new identifier
+		// new line with incorrect format, skip
+		if !ctx.headRegexp.Match(ctx.log) {
+			utils.Logger.Debug("log not match head regexp and there is no identifier exists",
+				zap.String("identifier", ctx.identifier),
+				zap.String("identifier_key", ctx.identifierKey),
+				zap.ByteString("log", ctx.log))
+			return msg
+		}
+
+		// new line with correct format, set as first line
+		utils.Logger.Debug("got new identifier",
+			zap.String("indentifier", ctx.identifier),
+			zap.ByteString("log", ctx.log))
+		ctx.identifier2LastMsg[ctx.identifier] = r.pendingMsgPool.Get().(*PendingMsg)
+		ctx.identifier2LastMsg[ctx.identifier].msg = msg
+		ctx.identifier2LastMsg[ctx.identifier].lastT = utils.Clock.GetUTCNow()
+		return nil
+	}
+
+	// replace exists msg in slot
+	if ctx.headRegexp.Match(ctx.log) { // new line
+		utils.Logger.Debug("got new line",
+			zap.ByteString("log", ctx.log),
+			zap.String("tag", msg.Tag))
+
+		newMsg = ctx.identifier2LastMsg[ctx.identifier].msg
+		ctx.identifier2LastMsg[ctx.identifier].msg = msg
+		ctx.identifier2LastMsg[ctx.identifier].lastT = utils.Clock.GetUTCNow()
+		return newMsg
+	}
+
+	// need to concat
+	utils.Logger.Debug("concat lines",
+		zap.String("tag", msg.Tag),
+		zap.ByteString("log", msg.Message[ctx.msgKey].([]byte)))
+	// ctx.identifier2LastMsg[ctx.identifier].msg.Message[ctx.msgKey] =
+	// 	append(ctx.identifier2LastMsg[ctx.identifier].msg.Message[ctx.msgKey].([]byte), '\n')
+	ctx.identifier2LastMsg[ctx.identifier].msg.Message[ctx.msgKey] =
+		append(ctx.identifier2LastMsg[ctx.identifier].msg.Message[ctx.msgKey].([]byte), msg.Message[ctx.msgKey].([]byte)...)
+
+	// too long to send
+	if len(ctx.identifier2LastMsg[ctx.identifier].msg.Message[ctx.msgKey].([]byte)) >= r.ConcatMaxLen {
+		utils.Logger.Debug("too long to send", zap.String("", ctx.msgKey), zap.String("tag", msg.Tag))
+		newMsg = ctx.identifier2LastMsg[ctx.identifier].msg
+		r.pendingMsgPool.Put(ctx.identifier2LastMsg[ctx.identifier])
+		delete(ctx.identifier2LastMsg, ctx.identifier)
+		return newMsg
+	}
+
+	// discard concated msg
+	r.msgPool.Put(msg)
+	return nil
 }