+++ tagFilters/base.go
@@ -8,7 +8,7 @@ import (
 	"github.com/Laisky/go-utils"
 	"github.com/Laisky/zap"
 	"github.com/cespare/xxhash"
-	"github.com/json-iterator/go"
+	jsoniter "github.com/json-iterator/go"
 )
 
 var json = jsoniter.ConfigCompatibleWithStandardLibrary
@@ -19,14 +19,14 @@ type TagFilterFactoryItf interface {
 	GetName() string
 
 	SetMsgPool(*sync.Pool)
-	SetCommittedChan(chan<- int64)
+	SetCommittedChan(chan<- *libs.FluentMsg)
 	SetDefaultIntervalChanSize(int)
 	DiscardMsg(*libs.FluentMsg)
 }
 
 type BaseTagFilterFactory struct {
 	msgPool                 *sync.Pool
-	committedChan           chan<- int64
+	committedChan           chan<- *libs.FluentMsg
 	defaultInternalChanSize int
 }
 
@@ -34,7 +34,7 @@ func (f *BaseTagFilterFactory) SetMsgPool(msgPool *sync.Pool) {
 	f.msgPool = msgPool
 }
 
-func (f *BaseTagFilterFactory) SetCommittedChan(committedChan chan<- int64) {
+func (f *BaseTagFilterFactory) SetCommittedChan(committedChan chan<- *libs.FluentMsg) {
 	f.committedChan = committedChan
 }
 
@@ -43,15 +43,7 @@ func (f *BaseTagFilterFactory) SetDefaultIntervalChanSize(size int) {
 }
 
 func (f *BaseTagFilterFactory) DiscardMsg(msg *libs.FluentMsg) {
-	f.committedChan <- msg.Id
-	if msg.ExtIds != nil {
-		for _, id := range msg.ExtIds {
-			f.committedChan <- id
-		}
-	}
-
-	msg.ExtIds = nil
-	f.msgPool.Put(msg)
+	f.committedChan <- msg
 }
 
 func (f *BaseTagFilterFactory) runLB(lbkey string, inChan chan *libs.FluentMsg, inchans []chan *libs.FluentMsg) {